fix(prism-registry): include_str hub py templates for loc-cap
